+++ tripy/tests/integration/test_layernorm.py
@@ -51,7 +51,7 @@ def setup(dtype, input_shape, normalized_shape):
     )
 
     input = torch.empty(*input_shape, dtype=torch_dtype, device="cuda").uniform_(0, 10)
-    tp_input = tp.Tensor(input, dtype=dtype)
+    tp_input = tp.Tensor(input)
 
     yield layernorm, tp_layernorm, tp_input
